requirements.txt
3-1==1.0.0
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-appconf==1.0.4
django-cleanup==5.1.0
django-imagekit==4.0.2
django-pwa==1.0.10
django-tinymce==3.2.0
django-widget-tweaks==1.4.8
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
pilkit==2.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
